Extract the Recaptcha code into a separate module
